requirements.txt
PyZenity==0.1.4
argparse==1.2.1
httplib2==0.8
oauth2==1.5.211
python-twitter==1.0
simplejson==3.3.0
wsgiref==0.1.2
